// server/src/database/interfaces.d.ts
export namespace std {
  export interface BaseObject {
    "id": string;
  }
  export interface $Object extends BaseObject {}
  export interface FreeObject extends BaseObject {}
  export type JsonEmpty = "ReturnEmpty" | "ReturnTarget" | "Error" | "UseNull" | "DeleteKey";
  export namespace enc {
    export type Base64Alphabet = "standard" | "urlsafe";
  }
}
// ...
export interface Region extends std.$Object {
  "open"?: boolean | null;
  "posts"?: unknown | null;
  "x": number;
  "y": number;
  "odds"?: number | null;
  "key": string;
  "digs"?: unknown | null;
}